% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccsICD9.R
\docType{data}
\name{ccsICD9}
\alias{ccsICD9}
\title{ccsICD9 data}
\format{A data frame with variables:
\describe{
\item{ICD}{ICD-9 short format (without decimal)}
\item{CCSCate}{CCS category}
\item{CCSDesc}{CCS name}
\item{ICDDesc}{ICD-9 name}
\item{OptCCSCate}{Optional CCS category}
\item{OptCCSDesc}{Optional CCS name}
\item{ICDD}{ICD-9 long format (with decimal)}
}}
\source{
\url{https://www.hcup-us.ahrq.gov/toolssoftware/ccs/ccs.jsp}
}
\usage{
ccsICD9
}
\description{
The Clinical Classifications Software (CCS) for ICD-9-CM is a
diagnosis and procedure categorization scheme that can be
employed in many types of projects analyzing data on diagnoses
and procedures. CCS is based on the International Classification
of Diseases, 9th Revision, Clinical Modification (ICD-9-CM),
a uniform and standardized coding system.
}
\keyword{datasets}

