% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDiagPeriod.R
\name{getDiagPeriod}
\alias{getDiagPeriod}
\title{Get the first diagnosis record}
\usage{
getDiagPeriod(caseAllIcd, caseFirstIcd)
}
\description{
This can be used to select the first diagnosis record
based on ICD code (grepl language) in factIcd Data,
return first diagnosis record based on factIcd Data
}
\examples{
getDiagPeriod(caseAllIcd,caseFirstIcd)

}

