% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDrugEra.R
\name{getDrugEra}
\alias{getDrugEra}
\title{Generate Drug Era}
\usage{
getDrugEra(CasePharm, window = 30)
}
\arguments{
\item{CasePharm}{data.frame with MemberID (chr),Desc (chr),DispenseDate (Date),DaysSupply (num)}

\item{window}{allowed gap between pharmacy claims, default is 30}
}
\description{
This can be used to merge pharmacy claims data into drug era with defined window. Exposure days will be cacluated, too.
}
\examples{
getDrugEra(CasePharm) replace CasePharm to your own dataset

}

