% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFirstDiagBasedOnFactIcdList.R
\name{getFirstDiagBasedOnFactIcdList}
\alias{getFirstDiagBasedOnFactIcdList}
\title{Get the first diagnosis record}
\usage{
getFirstDiagBasedOnFactIcdList(greplString, factIcd)
}
\arguments{
\item{greplString}{ICD selection rules with grepl expression}

\item{factIcd}{ICD dataset with at least 3 columns: "MemberID","Date","ICD"}
}
\description{
This can be used to select the first diagnosis record
based on ICD code (grepl language) in factIcd Data,
return first diagnosis record based on factIcd Data
}
\examples{
getFirstDiagBasedOnFactIcdList('255|256',sampleIcdData)

}

