% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getICDName.R
\name{getICDName}
\alias{getICDName}
\title{Get the first diagnosis record}
\usage{
getICDName(icdList)
}
\description{
This can be used to select the first diagnosis record
based on ICD code (grepl language) in factIcd Data,
return first diagnosis record based on factIcd Data
}
\examples{
getICDName(icdList)

}

