% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupICDBasedOnCCS.R
\name{groupICDBasedOnCCS}
\alias{groupICDBasedOnCCS}
\title{Get the first diagnosis record}
\usage{
groupICDBasedOnCCS(icdList)
}
\description{
This can be used to select the first diagnosis record
based on ICD code (grepl language) in factIcd Data,
return first diagnosis record based on factIcd Data
}
\examples{
groupICDBasedOnCCS(icdList)

}

