% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectCasesBasedOnFactIcdList.R
\name{selectCasesBasedOnFactIcdList}
\alias{selectCasesBasedOnFactIcdList}
\title{Select cases based on ICD code and the number of ICD codes}
\usage{
selectCasesBasedOnFactIcdList(greplString, factIcd, ICDNumber)
}
\arguments{
\item{greplString}{ICD selection rules with grepl expression}

\item{factIcd}{ICD dataset with at least 3 columns: "MemberID","Date","ICD"}

\item{ICDNumber}{a threshold of number of ICD for case selection}
}
\description{
This can be used to select qualified cases from factIcd data
based on the ICD code searching criteria and number of ICD code
per patients in the inout factIcd dataset. Return MemberID
}
\examples{
selectCasesBasedOnFactIcdList('255|256',sampleIcdData,2)

}

