% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDrugEra.R
\name{getDrugEra}
\alias{getDrugEra}
\title{Caculate Drug Era}
\usage{
getDrugEra(CasePharm)
}
\arguments{
\item{CasePharm}{data.frame with MemberID (chr),Desc (chr),DispenseDate (Date),DaysSupply (num)}
}
\description{
Caculate Drug Era
}
\examples{
getDrugEra()

}

